refactor: rename files to describe types handled
