+++ src/compiler/astToDag.ts
